+++ packages/ts-newskit/src/slices/lead-story-3/article-stack.tsx
@@ -18,7 +18,7 @@ export const ArticleStack = ({
 }: ArticlesProps) => {
   const modifiedArticles = leadArticles.map(leadArticle => ({
     ...leadArticle,
-    imageTop: true
+    hideImage: true
   }));
 
   return (