+++ packages/article/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.16.10](https://github.com/newsuk/times-components/compare/@times-components/article@7.16.9...@times-components/article@7.16.10) (2023-07-25)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.16.9](https://github.com/newsuk/times-components/compare/@times-components/article@7.16.8...@times-components/article@7.16.9) (2023-07-25)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.16.8](https://github.com/newsuk/times-components/compare/@times-components/article@7.16.7...@times-components/article@7.16.8) (2023-07-25)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.16.7](https://github.com/newsuk/times-components/compare/@times-components/article@7.16.6...@times-components/article@7.16.7) (2023-07-24)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.16.6](https://github.com/newsuk/times-components/compare/@times-components/article@7.16.5...@times-components/article@7.16.6) (2023-07-24)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.16.5](https://github.com/newsuk/times-components/compare/@times-components/article@7.16.4...@times-components/article@7.16.5) (2023-07-24)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
 ## [7.16.4](https://github.com/newsuk/times-components/compare/@times-components/article@7.16.3...@times-components/article@7.16.4) (2023-07-21)
 
 **Note:** Version bump only for package @times-components/article