+++ packages/gradient/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.9](https://github.com/newsuk/times-components/compare/@times-components/gradient@3.5.8...@times-components/gradient@3.5.9) (2023-07-25)
+
+**Note:** Version bump only for package @times-components/gradient
+
+
+
+
+
+## [3.5.8](https://github.com/newsuk/times-components/compare/@times-components/gradient@3.5.7...@times-components/gradient@3.5.8) (2023-07-25)
+
+**Note:** Version bump only for package @times-components/gradient
+
+
+
+
+
 ## [3.5.7](https://github.com/newsuk/times-components/compare/@times-components/gradient@3.5.6...@times-components/gradient@3.5.7) (2023-04-19)
 
 **Note:** Version bump only for package @times-components/gradient