+++ packages/tracking/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.20.1](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.20.0...@times-components/tracking@2.20.1) (2023-07-25)
+
+
+### Bug Fixes
+
+* **TMRX-1314:** fix image jumping issue ([#3450](https://github.com/newsuk/times-components/issues/3450)) ([8b19feb](https://github.com/newsuk/times-components/commit/8b19feb5c74e115a6b0e7bef61d0efc6e0535a8b))
+
+
+
+
+
 # [2.20.0](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.19.4...@times-components/tracking@2.20.0) (2023-04-19)